Added CSS and Bootstrap to html templates.
Web page templates updated to use Bootstrap to
add mobile functionality, and CSS to add a
semblance of UI to the front end.

Small tweaks to the actual layout of the pages
to complement the changes were also added. Notably,
a line of text that clashed with the overal design
of the web pages was removed, and a unit test had
to be changed to facilitate the template change.